Finition des interfaces Creation de l'interface pour la vitesse

+++ TP LO12/TP1/Texte.h
@@ -7,13 +7,13 @@
 class Texte
 {
 public:
-	Texte(std::string str, GLdouble x = 0, GLdouble y = 0);
+	Texte(std::string str, GLfloat x = 0, GLfloat y = 0);
 	virtual ~Texte();
 	void affiche();
 	
 	void setTexte(std::string str);
 	void setPolice(void* font);
-	void setPosition(GLdouble x, GLdouble y);
+	void setPosition(GLfloat x, GLfloat y);
 	void setCouleur(GLfloat r, GLfloat g, GLfloat b);
 private:
 	void beginTexte();
@@ -23,8 +23,8 @@ private:
 private:
 	void* _police;
 	std::string _texte;
-	GLdouble _x;
-	GLdouble _y;
+	GLfloat _x;
+	GLfloat _y;
 	Vector3d<GLfloat> _couleur;
 };
